Math/SunEphemeris: use int_fast32_t instead of "long"

+++ src/Math/SunEphemeris.cpp
@@ -11,6 +11,8 @@
 #include "time/BrokenDateTime.hpp"
 #include "time/RoughTime.hpp"
 
+#include <cstdint>
+
 /** Sun radius in degrees (?) */
 static constexpr double SUN_DIAMETER = 0.53;
 
@@ -33,9 +35,9 @@ FNday(const BrokenDateTime date_time) noexcept
 {
   assert(date_time.IsPlausible());
 
-  long int luku = -7 * (date_time.year + (date_time.month + 9) / 12) / 4 +
+  int_fast32_t luku = -7 * (date_time.year + (date_time.month + 9) / 12) / 4 +
                   275 * date_time.month / 9 + date_time.day +
-                  (long int)date_time.year * 367;
+    static_cast<int_fast32_t>(date_time.year) * 367;
 
   return double(luku) - 730531.5 + (date_time.hour % 24) / 24.;
 }
